♻️(front) reorganize filter types to enable easier reuse
- Move filter group names from the reducer to a more logical place
  in the types folder
- Rename FilterDefinition to filters as it contains more than just
  that one interface
- Updates types wherever applicable to use string literals instead
  of plain strings
